requirements.txt
# requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.5
alembic==1.10.2